before channels

+++ accounts/views.py
@@ -6,9 +6,10 @@ from djoser.social.views import ProviderAuthView
 from rest_framework_simplejwt.views import(TokenObtainPairView, TokenRefreshView, TokenVerifyView)
 from .serializers import CustomTokenObtainPairSerializer
 
-from rest_framework import generics, permissions, status
+from rest_framework import generics, permissions
 from rest_framework.response import Response
-from .models import LandlordProfile, TenantProfile
+from rest_framework import status as drf_status
+from .models import LandlordProfile, TenantProfile, PricingTier
 from .serializers import LandlordProfileSerializer, TenantProfileSerializer
 
 
@@ -94,7 +95,7 @@ class CustomTokenVerifyView(TokenVerifyView):
 
 class LogoutView(APIView):
     def post(self, request, *args, **kwargs):
-        response = Response(status=status.HTTP_204_NO_CONTENT)
+        response = Response(status=drf_status.HTTP_204_NO_CONTENT)
         response.delete_cookie(settings.AUTH_COOKIE)
         response.delete_cookie('refresh')
         return response
@@ -108,10 +109,21 @@ class LandlordProfileView(generics.RetrieveUpdateAPIView):
     permission_classes = [permissions.IsAuthenticated]
 
     def get_object(self):
+        if self.request.user.user_type != 'landlord':
+            return None
         return LandlordProfile.objects.get_or_create(user=self.request.user)[0]
+
+    def retrieve(self, request, *args, **kwargs):
+        instance = self.get_object()
+        if instance is None:
+            return Response(status=drf_status.HTTP_403_FORBIDDEN)
+        serializer = self.get_serializer(instance)
+        return Response(serializer.data)
 
     def patch(self, request, *args, **kwargs):
         instance = self.get_object()
+        if instance is None:
+            return Response(status=drf_status.HTTP_403_FORBIDDEN)
         serializer = self.get_serializer(instance, data=request.data, partial=True)
         serializer.is_valid(raise_exception=True)
         serializer.save()
@@ -119,18 +131,31 @@ class LandlordProfileView(generics.RetrieveUpdateAPIView):
 
     def delete(self, request, *args, **kwargs):
         instance = self.get_object()
+        if instance is None:
+            return Response(status=drf_status.HTTP_403_FORBIDDEN)
         instance.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        return Response(status=drf_status.HTTP_204_NO_CONTENT)
 
 class TenantProfileView(generics.RetrieveUpdateAPIView):
     serializer_class = TenantProfileSerializer
     permission_classes = [permissions.IsAuthenticated]
 
     def get_object(self):
+        if self.request.user.user_type != 'tenant':
+            return None
         return TenantProfile.objects.get_or_create(user=self.request.user)[0]
+
+    def retrieve(self, request, *args, **kwargs):
+        instance = self.get_object()
+        if instance is None:
+            return Response(status=drf_status.HTTP_403_FORBIDDEN)
+        serializer = self.get_serializer(instance)
+        return Response(serializer.data)
 
     def patch(self, request, *args, **kwargs):
         instance = self.get_object()
+        if instance is None:
+            return Response(status=drf_status.HTTP_403_FORBIDDEN)
         serializer = self.get_serializer(instance, data=request.data, partial=True)
         serializer.is_valid(raise_exception=True)
         serializer.save()
@@ -138,8 +163,10 @@ class TenantProfileView(generics.RetrieveUpdateAPIView):
 
     def delete(self, request, *args, **kwargs):
         instance = self.get_object()
+        if instance is None:
+            return Response(status=drf_status.HTTP_403_FORBIDDEN)
         instance.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        return Response(status=drf_status.HTTP_204_NO_CONTENT)
     
     
 # class LandlordListView(generics.ListAPIView):
@@ -175,4 +202,204 @@ class TenantProfileListView(generics.ListAPIView):
     
             
         
+from paynow import Paynow
+from django.conf import settings
+from .models import Payment
+import hashlib
+from rest_framework.permissions import AllowAny
+import uuid
+import logging
+import time
+logger = logging.getLogger(__name__)
+class InitiatePaymentView(APIView):
+    # permission_classes = [AllowAny]
 
+    def post(self, request):
+        tenant_profile = TenantProfile.objects.get(user=request.user)
+        email = request.data.get('email')
+        phone = request.data.get('phone')
+        # amount = request.data.get('amount')
+        plan = request.data.get('plan')
+        if plan == 'basic':
+            amount = 5
+            pricing_tier = PricingTier.objects.get(name='Basic')
+        elif plan == 'standard':
+            amount = 15
+            pricing_tier = PricingTier.objects.get(name='Standard')
+        elif plan == 'premium':
+            amount = 30
+            pricing_tier = PricingTier.objects.get(name='Premium')
+        elif plan == 'luxury':
+            amount = 50
+            pricing_tier = PricingTier.objects.get(name='Luxury')
+        else:
+            return Response({'error': 'Invalid plan'}, status=drf_status.HTTP_400_BAD_REQUEST)
+
+        paynow = Paynow(
+            settings.PAYNOW_INTEGRATION_ID,
+            settings.PAYNOW_INTEGRATION_KEY,
+            settings.PAYNOW_RESULT_URL,
+            settings.PAYNOW_RETURN_URL
+        )
+
+        reference = f'Order_{uuid.uuid4()}'
+        payment = paynow.create_payment(reference, email)
+        payment.add('Order Payment', float(amount))
+
+        try:
+            response = paynow.send_mobile(payment, phone, 'ecocash')
+            
+            if response.success:
+                time.sleep(30)
+                print('success')
+                status = paynow.check_transaction_status(response.poll_url)
+                print(status.status)
+                Payment.objects.create(
+                    reference=reference,
+                    poll_url=response.poll_url,
+                    amount=amount,
+                    phone=phone,
+                    email=email,
+                    status = status.status,
+                    tenant = tenant_profile
+                )
+                # Update tenant profile
+                tenant_profile.subscription_plan = plan
+                tenant_profile.subscription_status = 'active'
+                tenant_profile.pricing_tier = pricing_tier
+                tenant_profile.num_properties = pricing_tier.max_properties
+                tenant_profile.save()
+                
+                return Response({
+                    'poll_url': response.poll_url,
+                    'reference': reference,
+                    'status': status.status
+                }, status=drf_status.HTTP_200_OK)
+            else:
+                return Response({'error': response.data['error']}, status=drf_status.HTTP_400_BAD_REQUEST)
+        
+        except Exception as e:
+            return Response({'error': str(e)}, status=drf_status.HTTP_500_INTERNAL_SERVER_ERROR)
+
+
+
+class PaymentResultView(APIView):
+    permission_classes = [AllowAny]  # Allow access without authentication
+    def post(self, request):
+
+        data = request.data
+        hash_string = '&'.join([f'{key}={data[key]}' for key in sorted(data.keys()) if key != 'hash'])
+        local_hash = hashlib.sha512((hash_string + settings.PAYNOW_INTEGRATION_KEY).encode('utf-8')).hexdigest()
+
+        if local_hash != data.get('hash'):
+            return Response({'error': 'Invalid hash'}, status=drf_status.HTTP_400_BAD_REQUEST)
+        status_data = request.data
+        reference = status_data.get('reference')
+        payment = Payment.objects.filter(reference=reference).first()
+
+        if payment:
+            payment.status = status_data.get('status')
+            payment.save()
+            return Response({'status': 'Updated'}, status=drf_status.HTTP_200_OK)
+        else:
+            return Response({'error': 'Payment not found'}, status=drf_status.HTTP_404_NOT_FOUND)
+        
+
+
+
+
+class PaymentStatusView(APIView):
+    permission_classes = [AllowAny]  # Allow access without authentication
+
+    def post(self, request):
+        poll_url = request.data.get('poll_url')
+        if not poll_url:
+            return Response({'error': 'Poll URL is required'}, status=drf_status.HTTP_400_BAD_REQUEST)
+
+        paynow = Paynow(
+            settings.PAYNOW_INTEGRATION_ID,
+            settings.PAYNOW_INTEGRATION_KEY,
+        )
+
+        response = paynow.poll_transaction(poll_url)
+
+        if response.success:
+            payment_status = response.status
+            # Update your payment model if needed
+            Payment.objects.filter(poll_url=poll_url).update(status=payment_status)
+            return Response({'status': payment_status}, status=drf_status.HTTP_200_OK)
+        else:
+            return Response({'error': 'Failed to get payment status'}, status=drf_status.HTTP_400_BAD_REQUEST)
+
+from rest_framework.views import APIView
+from rest_framework.response import Response
+from rest_framework import permissions, status
+from .models import TenantProfile
+from .serializers import TenantProfileLimitedSerializer
+
+class TenantProfileLimitedView(APIView):
+    permission_classes = [permissions.IsAuthenticated]
+
+    def get(self, request):
+        if request.user.user_type != 'tenant':
+            return Response({"error": "Only tenants can access this information."}, status=status.HTTP_403_FORBIDDEN)
+        
+        try:
+            tenant_profile = TenantProfile.objects.get(user=request.user)
+            serializer = TenantProfileLimitedSerializer(tenant_profile)
+            return Response(serializer.data)
+        except TenantProfile.DoesNotExist:
+            return Response({"error": "Tenant profile not found."}, status=status.HTTP_404_NOT_FOUND)
+        
+
+
+class LandlordProfileLimitedView(APIView):
+    permission_classes = [permissions.IsAuthenticated]
+
+    def get(self, request, pk):
+        if request.user.user_type != 'tenant':
+            return Response({"error": "Only tenants can access this information."}, status=status.HTTP_403_FORBIDDEN)
+        
+        try:
+            landlord_profile = LandlordProfile.objects.get(id=pk)
+            serializer = LandlordProfileSerializer(landlord_profile)
+            return Response(serializer.data)
+        except LandlordProfile.DoesNotExist:
+            return Response({"error": "Landlord profile not found."}, status=status.HTTP_404_NOT_FOUND)
+
+
+from rest_framework.views import APIView
+from rest_framework.response import Response
+from rest_framework import permissions, status
+from api.models import Property
+from .models import TenantProfile
+from django.db import transaction
+
+class AddTenantAccessView(APIView):
+    permission_classes = [permissions.IsAuthenticated]
+
+    @transaction.atomic
+    def post(self, request, property_id):
+        if request.user.user_type != 'tenant':
+            return Response({"error": "Only tenants can access properties."}, status=status.HTTP_403_FORBIDDEN)
+        
+        try:
+            property = Property.objects.get(id=property_id)
+            tenant_profile = TenantProfile.objects.get(user=request.user)
+
+            if tenant_profile.num_properties <= 0:
+                return Response({"error": "You have reached the maximum number of properties you can access."}, status=status.HTTP_400_BAD_REQUEST)
+
+            if request.user in property.tenants_with_access.all():
+                return Response({"error": "You already have access to this property."}, status=status.HTTP_400_BAD_REQUEST)
+
+            property.tenants_with_access.add(request.user)
+            tenant_profile.num_properties -= 1
+            tenant_profile.save()
+
+            return Response({"message": "Access granted to the property and number of accessible properties updated."}, status=status.HTTP_200_OK)
+
+        except Property.DoesNotExist:
+            return Response({"error": "Property not found."}, status=status.HTTP_404_NOT_FOUND)
+        except TenantProfile.DoesNotExist:
+            return Response({"error": "Tenant profile not found."}, status=status.HTTP_404_NOT_FOUND)
